+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Allow calling `move_on_after` and `cancel_after` without block
+
 ## 0.33 2020-03-08
 
 * Implement `Fiber#supervise` (WIP)